+++ src/modules/transaction/index.mjs
@@ -6,6 +6,7 @@ import PrismaProcessor from "@prisma-cms/prisma-processor";
 import chalk from "chalk";
 
 import Tx from 'ethereumjs-tx';
+import solc from 'solc';
 
 
 export class EthTransactionProcessor extends PrismaProcessor {
@@ -22,14 +23,18 @@ export class EthTransactionProcessor extends PrismaProcessor {
   async create(objectType, args, info) {
 
     const {
+      db,
       web3,
       currentUser,
     } = this.ctx;
 
     let {
       data: {
+        to,
+        contractSourceId,
         privateKey,
         type,
+        gasPrice,
         params = [],
         ...data
       },
@@ -45,6 +50,8 @@ export class EthTransactionProcessor extends PrismaProcessor {
       return this.addError("Необходимо авторизоваться");
     }
 
+    const chainId = await web3.eth.net.getId();
+
     let address;
 
 
@@ -56,6 +63,7 @@ export class EthTransactionProcessor extends PrismaProcessor {
     }
 
     const {
+      id: senderAccountId,
       address: from,
     } = ethAccount;
 
@@ -90,7 +98,48 @@ export class EthTransactionProcessor extends PrismaProcessor {
       return this.addError("Аккаунт по приватному ключу и аккаунт пользователя не совпадают");
     }
 
-    params.unshift(from);
+
+
+    let Receiver;
+
+    if (to) {
+
+      if (to === from) {
+        return this.addError("Нельзя отправлять транзакции самому же себе");
+      }
+
+      const ReceiverAccount = await db.query.ethAccount({
+        where: {
+          address: to,
+        },
+      });
+
+      if (!ReceiverAccount) {
+        return this.addError("Не был найден получатель");
+      }
+
+
+      const {
+        id: receiverAccountId,
+        address: receiverAddress,
+      } = ReceiverAccount;
+
+
+      // params.unshift(receiverAddress);
+      Object.assign(args.data, {
+        to: receiverAddress,
+      });
+
+      Receiver = {
+        connect: {
+          id: receiverAccountId,
+        },
+      }
+
+    }
+
+
+    // params.unshift(from);
 
 
     switch (type) {
@@ -98,6 +147,11 @@ export class EthTransactionProcessor extends PrismaProcessor {
       case "SendEth":
 
         address = await this.sendEth(args, currentUserId);
+        break;
+
+      case "ContractCreate":
+
+        address = await this.createContract(args, currentUser, data);
         break;
 
     }
@@ -113,11 +167,18 @@ export class EthTransactionProcessor extends PrismaProcessor {
       args = {
         data: {
           ...data,
+          Receiver,
+          chainId,
           address,
           type,
-          CreatedBy: {
+          // CreatedBy: {
+          //   connect: {
+          //     id: currentUserId,
+          //   },
+          // },
+          Sender: {
             connect: {
-              id: currentUserId,
+              id: senderAccountId,
             },
           },
         },
@@ -167,36 +228,465 @@ export class EthTransactionProcessor extends PrismaProcessor {
 
     let {
       data: {
-        params: {
-          0: from,
-          1: to,
-          2: amount,
-        },
+        to,
+        gasPrice = 3,
+        amount,
+        params,
         privateKey,
       },
     } = args;
 
+    if (!to) {
+      return this.addError("Не был указан получатель");
+    }
 
-    privateKey = new Buffer(privateKey.replace(/^0x/, ''), 'hex');
+    if (!amount) {
+      return this.addFieldError("amount", "Не указана сумма перевода");
+    }
+    else if (amount < 0) {
+      return this.addFieldError("amount", "Сумма перевода не может быть отрицательной");
+    }
+
+    // const {
+    //   0: from,
+    //   1: to,
+    //   // 2: amount, 
+    // } = params;
+
+    let value = web3.utils.toWei(String(amount).replace(",", "."), 'ether');
+    value = web3.utils.toHex(value);
+
+    var rawTx = {
+      value,
+      to,
+      gasPrice: web3.utils.toHex(gasPrice * 10 ** 9),
+    };
+
+    return this.sendTransaction(rawTx, privateKey);
+  }
+
+
+  async createContract(args, currentUser, data) {
+
+    const {
+      id: currentUserId,
+      email,
+    } = currentUser;
+
+    let {
+      data: {
+        params,
+        privateKey,
+        contractSourceId,
+        gasPrice = 3,
+      },
+    } = args;
+
+
+    if (!contractSourceId) {
+      return this.addError("Не был указан ID кода контракта");
+    }
+
+
+    const {
+      db,
+      web3,
+    } = this.ctx;
+
+
+    const chainId = await web3.eth.net.getId();
+
+
+    const contractSource = await this.query("ethContractSource", {
+      where: {
+        id: contractSourceId,
+      },
+    }, `{
+      id
+      name
+      source
+      CreatedBy{
+        id
+      }
+    }`);
+
+    const {
+      name,
+      source,
+      CreatedBy: {
+        id: createdById,
+      },
+    } = contractSource;
+
+
+    // Контракт, который должен быть в результате создан
+    let Account;
+
+
+    // console.log("contractSource", contractSource);
+
+    // return;
+
+
+    // if (!createdById || createdById !== currentUserId) {
+    //   this.addError("Нельзя деплоить чужой контракт");
+    // }
+
+    // else if (!from) {
+    //   this.addError("Не был получен кошелек пользователя");
+    // }
+
+    // else 
+    if (!source) {
+      this.addError("Не был получен исходный код контракта");
+    }
+
+    // else if (!password) {
+    //   this.addFieldError("password", "Не указан пароль для кошелька");
+    // }
+
+    else {
+
+
+      let contractSourcesSource;
+
+
+      try {
+        contractSourcesSource = solc.compile(source);
+      }
+      catch (error) {
+
+
+        throw (error);
+      }
+
+
+      // console.log("contractSourcesSource", contractSourcesSource);
+
+      let contractSources = [];
+
+      if (contractSourcesSource) {
+
+        const {
+          errors,
+          contracts,
+        } = contractSourcesSource;
+
+        if (errors && errors.length) {
+          // throw new Error(errors);
+          console.error(chalk.cyan("Compele errors"), errors);
+        }
+
+        // for (var contractSourceName in contractSourcesSource.contractSources) {
+        for (var contractSourceName in contracts) {
+
+          // const contractSource = contractSourcesSource.contractSources[contractSourceName];
+          const contractSource = contracts[contractSourceName];
+
+          let {
+            interface: contractSourceInterface,
+          } = contractSource;
+
+          if (contractSourceInterface) {
+
+            try {
+
+              contractSourceInterface = JSON.parse(contractSourceInterface);
+
+            }
+            catch (e) {
+              console.error("Compele contract error", e);
+              throw (e);
+            }
+
+          }
+
+          // code and ABI that are needed by web3
+
+
+          // console.log("contractSources name", contractSourceName.replace(/^:*/, ''), name);
+
+
+          contractSources.push({
+            ...contractSource,
+            name: contractSourceName.replace(/^:*/, ''),
+            abi: contractSourceInterface,
+          });
+
+        }
+      }
+      else {
+        throw new Error("Не удалось скомпиллировать контракт");
+      }
+
+      // console.log("contractSources", contractSources);
+
+      let contractSourceForDeploy = contractSources.find(n => n.name === name);
+
+
+
+
+
+
+      // return;
+
+      if (!contractSourceForDeploy) {
+        this.addFieldError("name", "Не был получен публикуемый контракт");
+      }
+      else {
+
+        /**
+         * Если скомпиллированный контракт был получен,
+         * отправляем транзакцию,
+         * а исходники контракта добавляем к аккаунту
+         */
+
+        const {
+          abi,
+          bytecode,
+        } = contractSourceForDeploy;
+
+
+        Account = {
+          name,
+          chainId,
+          abi,
+          bytecode,
+          source,
+          type: "Contract",
+          ContractSource: {
+            connect: {
+              id: contractSourceId,
+            },
+          },
+          CreatedBy: {
+            connect: {
+              id: currentUserId,
+            },
+          },
+        };
+
+
+        // await this.ethDeployContractSource({
+        //   abi,
+        //   bytecode,
+        //   from,
+        //   password,
+        // })
+        //   .then(deployResult => {
+
+
+
+        //     const {
+        //       newContractSourceInstance,
+        //       txHash,
+        //     } = deployResult;
+
+
+
+
+        //     const {
+        //       _jsonInterface,
+        //       _address,
+        //     } = newContractSourceInstance;
+
+        //     Object.assign(data, {
+        //       Deployed: {
+        //         create: {
+        //           name,
+        //           source,
+        //           address: _address,
+        //           txHash,
+        //           bytecode,
+        //           abi: _jsonInterface,
+        //           CreatedBy: {
+        //             connect: {
+        //               id: currentUserId,
+        //             },
+        //           },
+        //         },
+        //       },
+        //     });
+
+        //   });
+
+
+      }
+
+    }
+
+
+    // args = {
+    //   where,
+    //   data: {
+    //     ...data
+    //   },
+    //   ...otherArgs
+    // };
+
+    // return this.sendTransaction(params, privateKey);
+
+    if (!Account) {
+      return this.addError("Не был сформирован аккаунт");
+    }
+
+    const {
+      abi,
+      bytecode,
+    } = Account;
+
+
+    // const {
+    //   address: contractAddress,
+    //   privateKey: contractPrivateKey,
+    // } = web3.eth.accounts.create(web3.utils.randomHex(32));;
+
+    // console.log("contract", contractAddress);
+
+
+    var contract = new web3.eth.Contract(abi);
+
+    // var contract = new web3.eth.Contract(abi, null, {
+    //   address: contractAddress,
+    // });
+
+    // contract.options.address = contractAddress;
+
+    // console.log("contract", contract.options);
+    // console.log("contract.address", contract.options.address);
+
+
+    // return;
+
+    let deploy = contract.deploy({
+      /**
+       * Если не передать hex, то будет ругаться, что не хватает газа
+       */
+      data: /^0x/.test(bytecode) ? bytecode : `0x${bytecode}`,
+      arguments: []
+    });
+
+    // console.log("contract deploy", deploy);
+
+    // return;
+
+    const address = await this.sendTransaction({
+      data: deploy.encodeABI(),
+      gasLimit: web3.utils.toHex(3000000),
+      gasPrice: web3.utils.toHex(gasPrice * 10 ** 9),
+
+      // Call contract method
+      // to: contract.options.address,
+      // toCreationAddress: true,
+    }, privateKey);
+
+    // console.log("contract address", address);
+
+
+    // const result = contract.deploy({
+    //   data: `0x${bytecode}`,
+    //   arguments: [],
+    // })
+
+    // console.log("contract result", result);
+
+
+    // return;
+
+
+    // let rawTx = {}
+
+
+    // console.log("rawTx", rawTx);
+
+    // return;
+
+    // // const address = await this.sendTransaction(params, privateKey);
+    // const address = await this.sendTransaction(rawTx, privateKey);
+
+    if (address) {
+
+
+
+      const txData = await web3.eth.getTransactionReceipt(address);
+
+      // console.log("txData", txData);
+
+      if (txData) {
+
+        const {
+          contractAddress,
+          status,
+        } = txData;
+
+        Object.assign(Account, {
+          address: contractAddress,
+        });
+
+
+        Object.assign(data, {
+          Account: {
+            create: Account,
+          },
+        });
+
+        // await db.mutation.createLetter({
+        //   data: {
+        //     rank: 100,
+        //     email,
+        //     subject: "Данные вашего контракта",
+        //     message: `
+        //     <h3>
+        //       Данные вашего контракта
+        //     </h3>
+  
+        //     <p>
+        //       <strong>Адрес:</strong> ${contractAddress}
+        //     </p>
+  
+        //     <p>
+        //       <strong>Приватный ключ:</strong> ${contractPrivateKey}
+        //     </p>
+        //   `,
+        //   }
+        // });
+
+      }
+
+
+    }
+
+    return address;
+  }
+
+
+  async sendTransaction(rawTx, privateKey) {
+
+    const {
+      web3,
+    } = this.ctx;
 
 
     if (this.hasErrors()) {
       return;
     }
 
-    let value = web3.utils.toWei(String(amount).replace(",", "."), 'ether');
-    value = web3.utils.toHex(value);
+    const {
+      address: from,
+    } = web3.eth.accounts.privateKeyToAccount(privateKey);
+
 
     let transactionCount = await web3.eth.getTransactionCount(from);
 
-    var rawTx = {
+
+    rawTx = {
       nonce: web3.utils.toHex(transactionCount),
-      value,
-      to,
-      gasLimit: web3.utils.toHex(21000),
-      gasPrice: web3.utils.toHex(3 * 10 ** 9),
+      gasLimit: web3.utils.toHex(210000),
+      gasPrice: web3.utils.toHex(4 * 10 ** 9),
+      ...rawTx,
     };
 
+    privateKey = new Buffer(privateKey.replace(/^0x/, ''), 'hex');
 
     var tx = new Tx(rawTx);
     tx.sign(privateKey);
@@ -205,9 +695,6 @@ export class EthTransactionProcessor extends PrismaProcessor {
 
 
     return new Promise((resolve, reject) => {
-
-      // let confirmationCount = 0;
-      // const minConfirmations = 3;
 
       let transactionHash;
 
@@ -233,15 +720,15 @@ export class EthTransactionProcessor extends PrismaProcessor {
 
         })
         .on('error', reject)
-        
-        /**
-         * Error: Failed to check for transaction receipt
-         * if use this
-         */
-        // .then(function (receipt) {
-        //   console.log("sendSignedTransaction then receipt");
-        //   return ;
-        // });
+
+      /**
+       * Error: Failed to check for transaction receipt
+       * if use this
+       */
+      // .then(function (receipt) {
+      //   console.log("sendSignedTransaction then receipt");
+      //   return ;
+      // });
 
     });